+++ lib/bali/path_helper.rb
@@ -7,12 +7,11 @@ module Bali
     # - exact: Matches the whole path
     # - partial: Matches if the path is included in the current path
     # - starts_with: Matches if the current path starts with the path
-    # - index_show: Matches the HREF by itself and with a number.
-    #               e.g. For /items, it matches /items/1 as well.
+    # - crud: Matches the HREF for any CRUD action.
     #
     # @param path [String] represents the path for the link
     # @param current_path [String] currently active URL
-    # @param match [Symbol] one of 4 match types [:exact, :partial, :starts_with, :index_show]
+    # @param match [Symbol] one of 4 match types [:exact, :partial, :starts_with, :crud]
     def active_path?(path, current_path, match: :exact)
       return false if current_path.nil?
 
@@ -20,9 +19,9 @@ module Bali
       current_request_path = current_path.gsub(/\.html$/, '')
 
       case match
-      when :index_show
+      when :crud
         path_without_params == current_request_path ||
-          path_with_number?(path, current_request_path)
+          show_or_edit_path?(path, current_request_path)
       when :starts_with
         current_request_path.starts_with?(path)
       when :partial
@@ -34,8 +33,8 @@ module Bali
 
     private
 
-    def path_with_number?(path, current_request_path)
-      %r{#{path}/[[:digit:]]+}.match(current_request_path)
+    def show_or_edit_path?(path, current_request_path)
+      %r{\A#{path}/[[:digit:]]+(/edit)?\Z}.match(current_request_path)
     end
   end
 end