+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.30.6] - 2022-07-18
 
-- And `index_show` match type to `SideMenu::Item`. So it only considers items as active when current_path equals the path or when it has an added number. E.g. When path is `items` ii will consider both `/items` and `/items/123` as active.
+- And `crud` match type to `SideMenu::Item`. So it only considers items as active when current_path is one of the CRUD actions (index, show, edit.)
 
 ## [0.30.5] - 2022-07-18